+++ vecep/src/components/layout/content/index.jsx
@@ -1,8 +1,8 @@
 import ContentContainer from './style.js';
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
-import Exercises from '../../../views/contents/exercises/Exercises';
-import Home from '../../../views/contents/home/Home.jsx';
+import Exercises from '../../../views/contents/exercises';
+import Home from '../../../views/contents/home';
 
 const Content = () => (
 	<ContentContainer>
