+++ vecep/src/views/App.js
@@ -1,8 +1,8 @@
 import './App.css';
 import React from 'react';
 import { BrowserRouter as Router} from 'react-router-dom';
-import Menu from '../components/layout/menu/Menu';
-import Content from '../components/layout/content/Content';
+import Menu from '../components/layout/menu/';
+import Content from '../components/layout/content/';
 
 const App = () => (
 	<div className="app">
