+++ vecep/src/views/contents/exercises/index.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useLocation } from 'react-router-dom';
-import Card from '../../../components/card/Card';
+import Card from '../../../components/card';
 import TextField from '@material-ui/core/TextField';
 import { ExercisesContainer, StyledAutocomplete } from './style';
 import { normalizeWord } from '../../../utils/normalizeWord';
